requirements.txt
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
# backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bleach==6.0.0
brotlipy==0.7.0
certifi==2022.12.7
cffi
chardet==4.0.0
charset-normalizer
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.0.7
cryptography
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
drf-writable-nested==0.7.0
executing==1.2.0
fastjsonschema==2.16.3
flatbuffers==23.3.3
flit_core
fonttools==4.39.0
fqdn==1.5.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.0.1
idna
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
# mkl-fft==1.3.1
# mkl-random
# mkl-service==2.4.0
mpmath==1.3.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy
oauthlib==3.2.2
omegaconf==2.3.0
onnx
onnxruntime
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pycparser
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL
pyparsing==3.0.9
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks
python-dateutil==2.8.2
python-json-logger==2.0.7
python-telegram-bot==20.1
python3-openid==3.2.0
pytorch-ood==0.1.0
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.25.1
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
scikit-learn==1.2.1
scipy==1.10.1
Send2Trash==1.8.0
six
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.11.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
--extra-index-url https://download.pytorch.org/whl/cu114
torch
torchvision
torchaudio
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions
tzdata==2022.7
uri-template==1.2.0
urllib3
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
win-inet-pton
wincertstore==0.2
zipp==3.15.0
